Move boundary operator assembly to batched/boundary.rs
